fixed: widget not refreshing and crashing.
For the not refreshing: the FirebaseRepo was converted into a singleton and it now adds valueListeners to each operation (wich must be removed after change if causes memory leakes) and calls the updateWidgetMethod.

For the crashing: On mulitple operations the widget getView method in the ViewsFactory(adapter) was troing a IndexOutOfBoundsException, this was solved with a try/catch and seems to be working fine. This seems to be happening due the sincronization with the firebaseDatabase (this has to be analyzed later.